requirements.txt
asgiref==3.7.2
Django==5.0.2
django-phonenumber-field==7.3.0
djangp-crispy-forms==2.4
crispy-tailwind==1.0.3
phonenumbers==8.13.30
sqlparse==0.4.4
tzdata==2024.1
python-dotenv==1.1.1
pillow==11.3.0